requirements.txt
dj-database-url==0.4.2
Django==1.10.5
django-braces==1.10.0
django-oauth-toolkit==0.11.0
django-taggit==0.21.3
djangorestframework==3.5.3
gunicorn==19.6.0
oauthlib==1.1.2
six==1.10.0
whitenoise==3.2.3
